requirements.txt
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.5
neo4j==5.20.0
streamlit==1.34.0
streamlit-chat==0.1.1
python-dotenv==1.0.1
flask
llama-index==0.9.48
llama-index-graph-stores-neo4j
llama-index-response-synthesizers-google
pinecone-client
neo4j-driver
sentence-transformers